requirements.txt
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
geographiclib==1.52
geopy==2.2.0
Pillow==9.0.1
